feat(rules): refactor validation command structure and update documentation
- Replaced the `uv run rules validate` command with `uv run ai validate-rules` for improved clarity and consistency in validation processes.
- Updated relevant documentation across multiple files to reflect the new command structure, ensuring developers have accurate guidance on validation procedures.
- Introduced a new `ai` command group to encapsulate AI-related tools and utilities, enhancing the organization of command functionalities.
- Removed the deprecated `rules.py` file, streamlining the codebase and focusing on the new validation approach.
